Avoid mount point check when cleaning mount points
